+++ myokit/tools.py
@@ -487,10 +487,12 @@ def natural_sort_key(s):
         for text in _natural_sort_regex.split(s)]
 
 
-def rmtree(path):
+def rmtree(path, silent=False):
     """
     Version of ``shutil.rmtree`` that handles Windows "access denied" errors
     (when the user is lacking write permissions, but is allowed to set them).
+
+    If ``silent=True`` any other exceptions will be caught and ignored.
     """
     # From https://stackoverflow.com/questions/2656322
     def onerror(function, path, excinfo):   # pragma: no cover
@@ -501,5 +503,11 @@ def rmtree(path):
         else:
             raise
 
-    shutil.rmtree(path, ignore_errors=False, onerror=onerror)
+    if silent:
+        try:
+            shutil.rmtree(path, ignore_errors=False, onerror=onerror)
+        except Exception:
+            pass
+    else:
+        shutil.rmtree(path, ignore_errors=False, onerror=onerror)